Fixed attempts to load directories
When selecting a directory node, we would try to load that as a file serverside. This would log a stack trace on the backend, but not disrupt the client side view. The serverside now prevents loading of directories, and instead returns an error. Client side will not submit post requests for nodes it knows are directories.